Fixed issue with displaying categories
If a category was chosen from the navbar/footer, before the Pictures.vue was loaded, any category would lead to the category 'normal'. Only the succeeding click would change the category appropriately.

The issue was fixed by adding an if-statement in Pictures.vue, that would reset the type once the feed was retrieved from the database.